+++ CHANGES.rst
@@ -149,6 +149,8 @@ Cubeviz
 Imviz
 ^^^^^
 
+- Virtual Observatory plugin to query resources and download data products. [#2872]
+
 Mosviz
 ^^^^^^